# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal_api.CreateLoadTestRequest:
    properties:
      accountId:
        type: string
      createdBy:
        type: string
      defaultDurationSec:
        type: integer
      defaultSpawnRate:
        type: number
      defaultUsers:
        type: integer
      description:
        type: string
      envId:
        type: string
      locustClusterId:
        type: string
      maxDurationSec:
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      orgId:
        type: string
      projectId:
        type: string
      scenarioId:
        type: string
      scriptContent:
        description: Base64 encoded Python script
        type: string
      tags:
        items:
          type: string
        type: array
      targetUrl:
        type: string
    required:
    - accountId
    - createdBy
    - locustClusterId
    - name
    - orgId
    - projectId
    - scriptContent
    - targetUrl
    type: object
  internal_api.CreateLoadTestRunRequest:
    properties:
      createdBy:
        type: string
      durationSeconds:
        description: Override from LoadTest
        type: integer
      loadTestId:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      spawnRate:
        description: Override from LoadTest
        type: number
      targetUsers:
        description: Override from LoadTest
        type: integer
    required:
    - createdBy
    - loadTestId
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  internal_api.LoadTestResponse:
    properties:
      accountId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      defaultDurationSec:
        type: integer
      defaultSpawnRate:
        type: number
      defaultUsers:
        type: integer
      description:
        type: string
      envId:
        type: string
      id:
        type: string
      latestRevisionId:
        type: string
      locustClusterId:
        type: string
      maxDurationSec:
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      orgId:
        type: string
      projectId:
        type: string
      scenarioId:
        type: string
      tags:
        items:
          type: string
        type: array
      targetUrl:
        type: string
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  internal_api.LoadTestRunResponse:
    properties:
      accountId:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      durationSeconds:
        type: integer
      envId:
        type: string
      finishedAt:
        type: string
      id:
        type: string
      lastMetrics:
        $ref: '#/definitions/internal_api.MetricSnapshotResponse'
      loadTestId:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      orgId:
        type: string
      projectId:
        type: string
      spawnRate:
        type: number
      startedAt:
        type: string
      status:
        type: string
      targetUsers:
        type: integer
      updatedAt:
        type: string
      updatedBy:
        type: string
    type: object
  internal_api.MetricSnapshotResponse:
    properties:
      avgResponseMs:
        type: number
      currentUsers:
        type: integer
      errorRate:
        type: number
      p50ResponseMs:
        type: number
      p95ResponseMs:
        type: number
      p99ResponseMs:
        type: number
      requestStats:
        additionalProperties:
          $ref: '#/definitions/internal_api.ReqStatResponse'
        type: object
      timestamp:
        type: string
      totalFailures:
        type: integer
      totalRequests:
        type: integer
      totalRps:
        type: number
    type: object
  internal_api.ReqStatResponse:
    properties:
      avgResponseTime:
        type: number
      maxResponseTime:
        type: number
      medianResponseTime:
        type: number
      method:
        type: string
      minResponseTime:
        type: number
      name:
        type: string
      numFailures:
        type: integer
      numRequests:
        type: integer
      requestsPerSec:
        type: number
    type: object
...
  internal_api.SuccessResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_api.UpdateLoadTestRequest:
    properties:
      defaultDurationSec:
        type: integer
      defaultSpawnRate:
        type: number
      defaultUsers:
        type: integer
      description:
        type: string
      maxDurationSec:
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      scenarioId:
        type: string
      tags:
        items:
          type: string
        type: array
      targetUrl:
        type: string
      updatedBy:
        type: string
    required:
    - updatedBy
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@loadmanager.io
    name: API Support
  description: |-
    Load testing management platform for creating, executing, and monitoring distributed load tests with Locust.

    ## Features
    - **Load Test Management**: Create and manage load tests with script versioning
    - **Script Revisions**: Full version control for test scripts with audit trail
    - **Test Execution**: Start and monitor load test runs with real-time metrics
    - **Visualization**: Real-time dashboards and historical metrics

    ## Authentication
    Currently, this API does not require authentication. Future versions will include API key or OAuth2 support.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Load Manager API
  version: "1.0"
paths:
  /load-tests:
    get:
      description: Returns a list of all load test configurations
      produces:
      - application/json
      responses:
        "200":
          description: List of load tests
          schema:
            items:
              $ref: '#/definitions/internal_api.LoadTestResponse'
            type: array
        "500":
          description: Failed to list load tests
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List all load tests
      tags:
      - LoadTests
    post:
      consumes:
      - application/json
      description: Creates a new load test configuration with an initial script revision
      parameters:
      - description: Load test configuration with base64 encoded script
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateLoadTestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Load test created successfully
          schema:
            $ref: '#/definitions/internal_api.LoadTestResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to create load test
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Create a new load test
      tags:
      - LoadTests
  /load-tests/{id}:
    delete:
      description: Deletes a load test configuration and all its associated data
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Load test deleted successfully
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "500":
          description: Failed to delete load test
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Delete a load test
      tags:
      - LoadTests
    get:
      description: Retrieves a specific load test configuration by its ID
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Load test details
          schema:
            $ref: '#/definitions/internal_api.LoadTestResponse'
        "404":
          description: Load test not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get load test by ID
      tags:
      - LoadTests
    put:
      consumes:
      - application/json
      description: Updates an existing load test configuration (excluding script)
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated load test configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateLoadTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Load test updated successfully
          schema:
            $ref: '#/definitions/internal_api.LoadTestResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Load test not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to update load test
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Update load test configuration
      tags:
      - LoadTests
  /load-tests/{id}/runs:
    post:
      consumes:
      - application/json
      description: Creates and starts a new load test run using the latest script
        revision
      parameters:
      - description: Load Test ID
        in: path
        name: id
        required: true
        type: string
      - description: Test run configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateLoadTestRunRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Load test run started successfully
          schema:
            $ref: '#/definitions/internal_api.LoadTestRunResponse'
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Load test not found or no script available
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to start load test run
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Start a new load test run
      tags:
      - Runs
...
  /runs:
    get:
      description: Returns a list of load test runs, optionally filtered by load test
        ID or other criteria
      parameters:
      - description: Load Test ID (when using /load-tests/{id}/runs endpoint)
        in: path
        name: id
        type: string
      - description: Filter by account ID
        in: query
        name: accountId
        type: string
      - description: Filter by organization ID
        in: query
        name: orgId
        type: string
      - description: Filter by project ID
        in: query
        name: projectId
        type: string
      - description: Filter by status (Pending, Running, Finished, Failed, Stopped)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of load test runs
          schema:
            items:
              $ref: '#/definitions/internal_api.LoadTestRunResponse'
            type: array
        "500":
          description: Failed to list load test runs
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: List load test runs
      tags:
      - Runs
  /runs/{id}:
    get:
      description: Retrieves details and current status of a specific load test run
      parameters:
      - description: Load Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Load test run details
          schema:
            $ref: '#/definitions/internal_api.LoadTestRunResponse'
        "404":
          description: Load test run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Get load test run details
      tags:
      - Runs
...
  /runs/{id}/stop:
    post:
      description: Stops a currently running load test run
      parameters:
      - description: Load Test Run ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Load test run stopped successfully
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "404":
          description: Load test run not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "500":
          description: Failed to stop load test run
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: Stop a running load test
      tags:
      - Runs
...
schemes:
- http
- https
swagger: "2.0"
...